% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce.R
\name{coalesce_tsd}
\alias{coalesce_tsd}
\alias{coalesce_tsd.TimeVectLayerVariable}
\alias{coalesce_tsd.TimeRastLayerVariable}
\title{coalesce TimeSpatData (Layer-Dimension)}
\usage{
coalesce_tsd(tsd_Data)

\method{coalesce_tsd}{TimeVectLayerVariable}(tsd_Data)

\method{coalesce_tsd}{TimeRastLayerVariable}(tsd_Data)
}
\arguments{
\item{tsd_Data}{\code{TimeSpatData} data
\itemize{
\item \code{TimeVectLayerVariable}
\item \code{TimeRastLayerVariable}
}}
}
\description{
Finds the first non-missing value at each position (Raster grid, Polygon, Line or Point) in Layer-Dimension.
\strong{NOTE}: this function can only be used for \code{TimeVectLayerVariable} and \code{TimeRastLayerVariable}.
}
