% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{crop_tsd}
\alias{crop_tsd}
\alias{crop_tsd.TimeVectVariable}
\alias{crop_tsd.TimeVectArray}
\alias{crop_tsd.TimeVectLayerVariable}
\alias{crop_tsd.TimeVectLayerArray}
\alias{crop_tsd.TimeRastVariable}
\alias{crop_tsd.TimeRastArray}
\alias{crop_tsd.TimeRastLayerVariable}
\alias{crop_tsd.TimeRastLayerArray}
\title{crop TimeSpatData (Spat-Dimension)}
\usage{
crop_tsd(tsd_Data, mask_Area)

\method{crop_tsd}{TimeVectVariable}(tsd_Data, mask_Area)

\method{crop_tsd}{TimeVectArray}(tsd_Data, mask_Area)

\method{crop_tsd}{TimeVectLayerVariable}(tsd_Data, mask_Area)

\method{crop_tsd}{TimeVectLayerArray}(tsd_Data, mask_Area)

\method{crop_tsd}{TimeRastVariable}(tsd_Data, mask_Area)

\method{crop_tsd}{TimeRastArray}(tsd_Data, mask_Area)

\method{crop_tsd}{TimeRastLayerVariable}(tsd_Data, mask_Area)

\method{crop_tsd}{TimeRastLayerArray}(tsd_Data, mask_Area)
}
\arguments{
\item{tsd_Data}{\code{TimeSpatData} data
\itemize{
\item \code{TimeVectVariable}
\item \code{TimeVectArray}
\item \code{TimeRastVariable}
\item \code{TimeRastArray}
\item \code{TimeRastLayerVariable}
\item \code{TimeRastLayerArray}
}}

\item{mask_Area}{(SpatVector) \strong{one} Polygon, from \code{\link[terra:vect]{terra::vect()}}}
}
\description{
Crop (mask) \code{TimeSpatData} new \code{TimeSpatData}.
Every time can only \strong{one Polygon} as the \code{mask_Area}.
\itemize{
\item for \code{TimeVectData} the polygons, lines oder points will mit \code{mask_Area} first intersect \link[terra:intersect]{terra::intersect}
then only the polygons, lines oder points which in the \code{mask_Area} will be exported.
\item for \code{TimeRastData} or \code{TimeRastLayerData} the spatial dimension will first crop \link[terra:crop]{terra::crop} than \link[terra:mask]{terra::mask}
the raster-cells will be in NA setted, when they are not in the \code{mask_Area}.
}

Crop (mask) \code{TimeSpatData} to create a new \code{TimeSpatData}.
Every time the function is used, \strong{only one polygon} can be used as the mask_Area.
\itemize{
\item For \code{TimeVectData}, the polygons, lines, or points will be intersected with the mask_Area using \link[terra:intersect]{terra::intersect},
and only the polygons, lines, or points within the \code{mask_Area} will be exported.
\item For \code{TimeRastData}, the spatial dimension will be cropped using \link[terra:crop]{terra::crop} and then masked using \link[terra:mask]{terra::mask}.
Raster cells that are not within the \code{mask_Area} will be set to NA.
}
}
