% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_tsd}
\alias{extract_tsd}
\alias{extract_tsd.TimeVectVariable}
\alias{extract_tsd.TimeVectArray}
\alias{extract_tsd.TimeRastVariable}
\alias{extract_tsd.TimeRastArray}
\title{extract TimeSpatData (Spat-Dimension)}
\usage{
extract_tsd(tsd_data, mask_area)

\method{extract_tsd}{TimeVectVariable}(tsd_data, mask_area)

\method{extract_tsd}{TimeVectArray}(tsd_data, mask_area)

\method{extract_tsd}{TimeRastVariable}(tsd_data, mask_area)

\method{extract_tsd}{TimeRastArray}(tsd_data, mask_area)
}
\arguments{
\item{tsd_data}{\code{TimeSpatData} data
\itemize{
\item \code{TimeVectVariable}
\item \code{TimeVectArray}
\item \code{TimeRastVariable}
\item \code{TimeRastArray}
}}

\item{mask_area}{(SpatVector) Polygons, from \code{\link[terra:vect]{terra::vect()}}}
}
\description{
Extract \code{TimeSpatData} to \code{TimeVectData}.
The theory is same as \link[terra:extract]{terra::extract}, \code{TimeVectData} or \code{TimeRastData} will be extracted (weighted average) in the given polygons (\code{mask_area}).
The weight will be calculated with the area of polygons or raster-cells, that in the \code{mask_area}.
\strong{NOTE}: this function can be used only for the \strong{value, that makes sense in weighted average}.
}
