% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_tsd}
\alias{read_tsd}
\title{read TimeSpatData}
\usage{
read_tsd(fn_, start = NA, count = NA)
}
\arguments{
\item{fn_}{(char) filenames}

\item{start, count}{(num) start and count od the data block, more see function \code{\link[ncdf4:ncvar_get]{ncdf4::ncvar_get()}}}
}
\value{
\code{TimeSpatData} data
\itemize{
\item TimeVectVariable
\item TimeVectArray
\item TimeRastVariable
\item TimeRastArray
\item TimeRastLayerVariable
\item TimeRastLayerArray
}
}
\description{
read \code{TimeSpatData} from nc-file, that writed by \code{\link[=write_tsd]{write_tsd()}},
or write by other tools that match the \code{TimeSpatData} convention.
}
