% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset}
\alias{subset}
\alias{[[.TimeVectArray}
\alias{[[.TimeRastArray}
\alias{[[.TimeRastLayerArray}
\title{read TimeSpatData}
\usage{
\method{[[}{TimeVectArray}(data_, name)

\method{[[}{TimeRastArray}(data_, name)

\method{[[}{TimeRastLayerArray}(data_, name)
}
\arguments{
\item{data_}{(TimeSpatArray) data in
\itemize{
\item \code{TimeVectArray},
\item \code{TimeRastArray} or
\item \code{TimeRastLayerArray}
}}

\item{name}{(char) name od the variable}
}
\value{
\code{TimeSpatVariable} data
}
\description{
get the variable from \code{TimeSpatArray} to \code{TimeSpatVariable}:
\itemize{
\item \code{TimeVectArray} to \code{TimeVectVariable}
\item \code{TimeRastArray} to \code{TimeRastVariable}
\item \code{TimeRastLayerArray} to \code{TimeRastLayerVariable}
}
}
