% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_summarise.R
\name{summary_time_tsd}
\alias{summary_time_tsd}
\alias{summary_index_tsd}
\title{summarize TimeSpatData in Time-Dimension}
\usage{
summary_time_tsd(tsd_Data, str_Interval = "years", smry_Funct = mean, ...)

summary_index_tsd(tsd_Data, idx_End, smry_Funct = mean, ...)
}
\arguments{
\item{tsd_Data}{\code{TimeSpatData} data
\itemize{
\item \code{TimeVectVariable}
\item \code{TimeVectArray}
\item \code{TimeRastVariable}
\item \code{TimeRastArray}
}}

\item{str_Interval}{(string) <“microseconds”, “milliseconds”, “seconds”, “minutes”, “hours”, “days”, “weeks”, “months”, “quarters” and “years”> time interval}

\item{smry_Funct}{(function) function that used for summary, e.g. \link{mean}, \link{max} and \link{min}, it can also be the self-defined function}

\item{...}{other parameters for \code{smry_Funct}}

\item{idx_End}{(integer) index of the end points every interval}
}
\value{
summarized time series with the same data structure like \code{tsd_Data}
}
\description{
\itemize{
\item \link{summary_time_tsd} summarize the Time-Dimension in fixed time interval like yearly, monthly and so on
\item \link{summary_index_tsd} summarize the Time-Dimension in flexibel time interval, use the parameter \code{idx_End} to divide the interval
}
}
