% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TimeRastData.R
\name{TimeRastArray}
\alias{TimeRastArray}
\alias{new_TimeRastArray}
\alias{new_TimeRastArray.SpatRaster}
\alias{new_TimeRastArray.array}
\title{create \code{TimeRastArray} data}
\usage{
new_TimeRastArray(data_, Name_, Unit_, Time_, ...)

\method{new_TimeRastArray}{SpatRaster}(data_, Name_, Unit_, Time_, dim_Time_Vari = c(1, 2), ...)

\method{new_TimeRastArray}{array}(data_, Name_, Unit_, Time_, Spat_EPSG, Spat_extent, ...)
}
\arguments{
\item{data_}{(num-array or terra::SpatRaster)
\itemize{
\item 4D (array) for \code{\link[=new_TimeRastArray.array]{new_TimeRastArray.array()}}, it must in the \strong{(time, x, y, variable)} dimension-order, use the function \code{\link[=aperm]{aperm()}} to adjust the dimension-order
\item (terra::SpatRaster) (multilayer) for \code{\link[=new_TimeRastArray.SpatRaster]{new_TimeRastArray.SpatRaster()}}, created by \code{\link[terra:rast]{terra::rast()}}
}}

\item{Name_, Unit_}{(char) name and unit of Variable, \code{Unit_} should be  \code{\link[units:units]{units::as_units()}}}

\item{Time_}{(vector of lubridate::timepoint) time dimension, created by \code{\link[lubridate:as_date]{lubridate::as_date()}} or \code{\link[lubridate:as_date]{lubridate::as_datetime()}}}

\item{...}{other parameters}

\item{dim_Time_Vari}{the order of time- and variable-dimenion in the layer of \code{terra::rast},
it must be
\itemize{
\item \code{c(1, 2)}: first time then variable or
\item \code{c(2, 1)}: first variable then time,
\item default is \code{c(1, 2)}
}}

\item{Spat_EPSG}{(integer) EPSG code of the CRS (Coordinate Reference System)}

\item{Spat_extent}{(vector of numric), extension of the Raster in c(xmin, xmax, ymin, ymax),
it can be projected east- and north-meter or global longitude- and latitude-degree}
}
\value{
\code{TimeRastArray} data
}
\description{
\code{TimeRastArray} is a data class that based on 4D-array. There are
time, spatial and variable three dimensions in "macro",
but actually due to the spatial dimension in 2D-raster,
so there are four dimension in data: \strong{(time, x, y, variable)}.
}
