% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TimeRastData.R
\name{TimeRastLayerArray}
\alias{TimeRastLayerArray}
\alias{new_TimeRastLayerArray}
\alias{new_TimeRastLayerArray.SpatRaster}
\alias{new_TimeRastLayerArray.array}
\title{create TimeRastLayerArray}
\usage{
new_TimeRastLayerArray(data_, Name_, Unit_, Time_, ...)

\method{new_TimeRastLayerArray}{SpatRaster}(
  data_,
  Name_,
  Unit_,
  Time_,
  n_Layer,
  dim_Time_Layer_Vari = c(1, 2, 3),
  ...
)

\method{new_TimeRastLayerArray}{array}(
  data_,
  Name_,
  Unit_,
  Time_,
  Spat_EPSG,
  Spat_extent,
  n_Layer,
  ...
)
}
\arguments{
\item{data_}{(num-array or terra::SpatRaster)
\itemize{
\item 3D (array) for \code{\link[=new_TimeRastVariable.array]{new_TimeRastVariable.array()}}, it must in the \strong{(time, x, y)} dimension-order, use the function \code{\link[=aperm]{aperm()}} to adjust the dimension-order
\item (terra::SpatRaster) (multilayer) for \code{\link[=new_TimeRastVariable.SpatRaster]{new_TimeRastVariable.SpatRaster()}}, created by \code{\link[terra:rast]{terra::rast()}}
}}

\item{Name_, Unit_}{(char) name and unit of Variable, \code{Unit_} should be  \code{\link[units:units]{units::as_units()}}}

\item{Time_}{(vector of lubridate::timepoint) time dimension, created by \code{\link[lubridate:as_date]{lubridate::as_date()}} or \code{\link[lubridate:as_date]{lubridate::as_datetime()}}}

\item{...}{other parameters}

\item{n_Layer}{(integer) number of the data layers}

\item{dim_Time_Layer_Vari}{the order of time- and variable-dimenion in the layer of \code{terra::rast},
it must be
\itemize{
\item \code{c(1, 2, 3)}, \code{c(1, 3, 2)}: first time then layer or variable,
\item \code{c(2, 1, 3)}, \code{c(2, 3, 1)}: first layer then time or variable,
\item \code{c(3, 1, 2)}, \code{c(3, 2, 1)}: first variable then time or layer,
\item default is \code{c(1, 2, 3)}
}}

\item{Spat_EPSG}{(integer) EPSG code of the CRS (Coordinate Reference System)}

\item{Spat_extent}{(vector of numric), extension of the Raster in c(xmin, xmax, ymin, ymax),
it can be projected east- and north-meter or global longitude- and latitude-degree}
}
\value{
\code{TimeRastLayerArray} data
}
\description{
\code{TimeRastLayerArray} have time, spatial two dimensions in "macro",
but actually due to the spatial dimension in 2D-raster, so there are there dimension in data.
}
