% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSpatData-package.R
\docType{package}
\name{TimeSpatData-package}
\alias{TimeSpatData-package}
\title{basic concept of TimeSpatData}
\description{
Die grundsätzliche Beschreibung:
}
\section{Zeitliche-räumliche Daten Strukturen definieren}{
\subsection{Makro Data-Dimensionen (\code{Dim}) definieren}{

In der Makrosicht ist das TimeSpatData in drei Dimensionen definiert:
\itemize{
\item \code{Time}: Zeitstampel definieren
\item \code{Spat}: Räumliche Struktur definieren
\item \code{Vari}: mehrere Variablen (Größe) definieren
}
}

\subsection{Datenstruktur (Array) Dimensionen (\code{dim}) definieren}{

Die Daten bei der Bearbeitung und Speicherung werden in Array-Form
gegeben, die Makro Dimensionen können nicht deutlich in eine
Array-Dimensionen definiert werden, deshalb die Array-Dimensionen sind
neu definiert:
\subsection{\code{time}}{

Dimension \code{time} ist für Zeit definiert und besitzt
\strong{Rückwärtskonsistenz}. Es steht immer in der ersten Dimension, mit
zwei Möglickeiten:
\itemize{
\item \code{continuous}: mit festem Zeitschritt, bei \code{units} und
\code{original_time} definiert
\item \code{discrete}: ohne dem Begriff Zeitschritt, sondern immer die genau
Zeitstempeln geben
}
}

\subsection{\code{spat}}{

Dimension \code{spat} ist für vector-Räumliche Daten (\code{points}, \code{lines}, und
\code{polygons}) definiert, es ist nicht ordnend aber ist es mit Varibale
\code{Spat_ID} und \code{Spat_Data} knüpft. Das CRS ist mit \strong{EPSG-code}
definiert.
}

\subsection{\code{x} und \code{y} in Horizontal}{

Die beide Dimensionen sind für Raster (einzeln Schichten) horizontal
Koordinate definiert mit die Koordinaten-Werten von zentralen Punkten
des Zells. Das CRS ist mit \strong{EPSG-code} definiert. Deshalb ist es nicht
erlaubt die Einheit oder andere zusätzliche Sachen von den beiden
Dimensionen zu definieren.
}

\subsection{\* \code{layer} und \code{z} in Vertikal}{

Die beiden Dimensionen sind für vertikale Schichten definiert.
\itemize{
\item \code{layer} meint eine konzeptionelle Schichten ohne genau vertikale
Lage
\item \code{z} mit genaue vertikale (vertikale CRS noch suchen), auch möglich
mit Reference Fläche:
\item oben von Reference Fläche
\item oder unten von Reference Fläche
}
}

\subsection{\code{vari}}{

Dimension \code{vari} ist für mehre entsprechende Größe (Variable) definiert
mit \code{Name} und \code{Unit}. Die Dimension steht immer am letzten.
}

}

\subsection{Daten Type definieren}{
\subsection{\code{Time}}{

Wegen den beiden kontinuierliche oder diskrete Form sind folgen beide
\code{Time-Form} definiert:
\itemize{
\item \code{ConTime [time]}: kontinuierliche Zeit
\item \code{DisTime [time]}: diskrete Zeit
}
}

\subsection{\code{Spat}}{

Wegen der Vielfalt von räumlichen Daten Form steht da unter fünf
\code{Spat-Form}:
\itemize{
\item \code{Vect [spat]}: Vector Daten (Punkten, Linien, Polygonen)
\item \code{VectLayer [spat, layer]}: Vector Daten mit mehrere Schichten
\item \code{Rast [x, y]}: Raster Daten in einzelner Schichte
\item \code{RastLayer [x, y, layer]}: Raster Daten mit mehrere Schichten
\item \code{RastVert [x, y, z]}: Raster Daten mit vertikale CRS
}
}

\subsection{\code{Vari}}{

Wegen der unterschiedlichen Ähnlichkeit zwischen Variable zu Variable,
in der Sicht kann Daten auch in drei \code{Vari-Form} definiert:
\itemize{
\item \verb{Vari [vari = 1] / []}: nur einzelne Variable
\item \code{Array [vari]}: mehrere Variablen gleiche Struktur und
Dimensiongroße besetzen
\item \code{Group [vari]}: mehrere Variablen gleiche Struktur und eine nicht
identische Dimensiongroße besetzen (meisten Zeit, besonders für
diskrete Zeitskala), aber diese flexibel Dimension gleichen Index
besetzen muss
}
}

\subsection{Time-Spat-Vari Daten}{

Zusammenfassend kann man sagen, ein komplettes TimeSpatData muss die
drei Makro Dimension (keine Daten Dimension) definieren. Es gibt
insgesamt 2 (Time) \* 5 (Spat) \* 3 (Vari) = 30 (DataForm). Aber nicht
alle Daten Type sind praktische, am folgen wird ein paar wichtige Daten
Formen gegebn:
\itemize{
\item \code{ConTimeVectVari}:
\item dim: [time, spat, vari] in 3D oder [time, spat] in 2D
\item z.B.: Durchfluss Zeitreihe in Pegel-Netz
\item \code{ConTimeVectArry}:
\item dim: [time, spat, vari] in 3D
\item z.B.: Meteorologische Zeitreihe (Temperatur, Luftfeuchte und
Windgeschwindigkeit) in Meteostation-Netz
\item \code{ConTimeRastArry}:
\item dim: [time, x, y, vari] in 4D
\item z.B.: reanalysieerte Meteorologische Zeitreihe (Temperatur,
Luftfeuchte und Windgeschwindigkeit) in Raster Form
}
}

}

\subsection{TimeSpatData in NetCDF}{

Die Definition von Dimensionen von NetCDF bleibt identisch wie oben
Datenstruktur Dimensionen. Wegen der Begrenze von NetCDF muss man auch
die andere zusätzliche Dimensionen, Variablen und Attributen weiter
definieren. Aus der Rücksicht auf Einheitligkeit und
Benutzerfreundlichkeit steht es folgende Konventionen:
\subsection{\code{Time}}{

Wegen den beiden kontinuierliche oder diskrete Form sind folgen beide
\code{Time-Form} definiert:
\itemize{
\item \code{ConTime [time]}: kontinuierliche Zeit
\item \code{DisTime [time]}: diskrete Zeit
}
}

}
}

