% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TimeVectData.R
\name{TimeVectData}
\alias{TimeVectData}
\alias{new_TimeVectVariable}
\alias{new_TimeVectArray}
\title{create \code{TimeVectVariable} and \code{TimeVectArray} data}
\usage{
new_TimeVectVariable(
  data_,
  Name_,
  Unit_,
  Time_,
  Spat_ID,
  Spat_Data,
  na_check = FALSE
)

new_TimeVectArray(
  data_,
  Name_,
  Unit_,
  Time_,
  Spat_ID,
  Spat_Data,
  na_check = FALSE
)
}
\arguments{
\item{data_}{(num-array) 2D for \code{TimeVectVariable} and 3D for \code{TimeVectArray}}

\item{Name_, Unit_}{(char or vector of char) name and unit of Variable, \code{Unit_} should be converted by \code{\link[units:units]{units::as_units()}}}

\item{Time_}{(vector of lubridate::timepoint) time dimension, created by \code{\link[lubridate:as_date]{lubridate::as_date()}} or \code{\link[lubridate:as_date]{lubridate::as_datetime()}}}

\item{Spat_ID}{(vector of char) the identifying of the spatial-dimension, they must be contained in the \code{Spat_Data}}

\item{Spat_Data}{(terra::SpatVector) geological data, create by \code{\link[terra:vect]{terra::vect()}}, it must contain the key (attribute) named as "Spat_ID"}

\item{na_check}{(bool) if check the NAs}
}
\description{
\itemize{
\item \code{TimeVectVariable} is a data class that is based on a 2D array. It has two dimensions: time and spatial.
The spatial dimension also holds the geological vector data (points, lines, polygons) that is saved in the \code{Spat_Data}.
\item \code{TimeVectArray} is a data class that is based on a 3D array. It has three dimensions: time, spatial, and variable.
The spatial dimension also holds the geological vector data (points, lines, polygons) that is saved in the \code{Spat_Data}.
}
}
