% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TimeVectLayerData.R
\name{TimeVectLayerData}
\alias{TimeVectLayerData}
\alias{new_TimeVectLayerVariable}
\alias{new_TimeVectLayerArray}
\title{create \code{TimeVectLayerVariable} and \code{TimeVectLayerArray} data}
\usage{
new_TimeVectLayerVariable(
  data_,
  Name_,
  Unit_,
  Time_,
  Spat_ID,
  Spat_Data,
  na_check = FALSE
)

new_TimeVectLayerArray(
  data_,
  Name_,
  Unit_,
  Time_,
  Spat_ID,
  Spat_Data,
  na_check = FALSE
)
}
\arguments{
\item{data_}{(num-array) 3D for \code{TimeVectLayerVariable} and 4D for \code{TimeVectLayerArray}}

\item{Name_, Unit_}{(char or vector of char) name and unit of Variable, \code{Unit_} should be converted by \code{\link[units:units]{units::as_units()}}}

\item{Time_}{(vector of lubridate::timepoint) time dimension, created by \code{\link[lubridate:as_date]{lubridate::as_date()}} or \code{\link[lubridate:as_date]{lubridate::as_datetime()}}}

\item{Spat_ID}{(vector of char) the identifying of the spatial-dimension, they must be contained in the \code{Spat_Data}}

\item{Spat_Data}{(terra::SpatVector) geological data, create by \code{\link[terra:vect]{terra::vect()}}}

\item{na_check}{(bool) if check the NAs}
}
\description{
\itemize{
\item \code{TimeVectLayerVariable} is a data class that is based on a 3D array. It has two "macro" dimensions: time and spatial.
The spatial dimension also holds the geological vector data (points, lines, polygons) that is saved in the \code{Spat_Data}.
But due to the vertical \strong{layer} in spatial dimension,
the data class actually has three data-dimensions: \strong{(time, spat, layer)}.
\item \code{TimeVectLayerArray} is a data class that is based on a 4D array. It has three dimensions: time, spatial, and variable.
The spatial dimension also holds the geological vector data (points, lines, polygons) that is saved in the \code{Spat_Data}.
But due to the vertical \strong{layer} in spatial dimension,
the data class actually has four data-dimensions: \strong{(time, spat, layer, variable)}.
}
}
