% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TimeVectData.R
\name{TimeVectVariable}
\alias{TimeVectVariable}
\alias{new_TimeVectVariable}
\alias{new_TimeVectArray}
\title{create \code{TimeVectVariable} data}
\usage{
new_TimeVectVariable(
  data_,
  Name_,
  Unit_,
  Time_,
  Spat_ID,
  Spat_Data,
  na_check = FALSE
)

new_TimeVectArray(
  data_,
  Name_,
  Unit_,
  Time_,
  Spat_ID,
  Spat_Data,
  na_check = FALSE
)
}
\arguments{
\item{data_}{(num-array) 3D numric array}

\item{Name_, Unit_}{(vector of char) names and units of Variables, \code{Unit_} should be converted by \code{\link[units:units]{units::as_units()}}}

\item{Time_}{(vector of lubridate::timepoint) time dimension, created by \code{\link[lubridate:as_date]{lubridate::as_date()}} or \code{\link[lubridate:as_date]{lubridate::as_datetime()}}}

\item{Spat_ID}{(vector of char) the identifying of the spatial-dimension, they must be contained in the \code{Spat_Data}}

\item{Spat_Data}{(terra::SpatVector) geological data, create by \code{\link[terra:vect]{terra::vect()}}}

\item{na_check}{(bool), if check the NAs}
}
\description{
\code{TimeVectVariable} is a data class based on 2D-array. There are
two dimensions: time- and spatial-dimension. The spatial-dimension holds also
the geological vector-data (points, lines, polygons), that is saved in the \code{Spat_Data}.

\code{TimeVectArray} is a data class based on 3D-array. There are
three dimensions: time-, spatial- and vraiable-dimension. The spatial-dimension holds also
the geological vector-data (points, lines, polygons), that is saved in the \code{Spat_Data}.
}
