% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_tsd}
\alias{write_tsd}
\alias{write_tsd.TimeVectVariable}
\alias{write_tsd.TimeVectArray}
\alias{write_tsd.TimeVectLayerVariable}
\alias{write_tsd.TimeVectLayerArray}
\alias{write_tsd.TimeRastVariable}
\alias{write_tsd.TimeRastArray}
\alias{write_tsd.TimeRastLayerVariable}
\alias{write_tsd.TimeRastLayerArray}
\title{write TimeSpatData}
\usage{
write_tsd(data_, fn_, other_Attr = NULL)

\method{write_tsd}{TimeVectVariable}(data_, fn_, other_Attr = NULL)

\method{write_tsd}{TimeVectArray}(data_, fn_, other_Attr = NULL)

\method{write_tsd}{TimeVectLayerVariable}(data_, fn_, other_Attr = NULL)

\method{write_tsd}{TimeVectLayerArray}(data_, fn_, other_Attr = NULL)

\method{write_tsd}{TimeRastVariable}(data_, fn_, other_Attr = NULL)

\method{write_tsd}{TimeRastArray}(data_, fn_, other_Attr = NULL)

\method{write_tsd}{TimeRastLayerVariable}(data_, fn_, other_Attr = NULL)

\method{write_tsd}{TimeRastLayerArray}(data_, fn_, other_Attr = NULL)
}
\arguments{
\item{data_}{\code{TimeSpatData} data
\itemize{
\item \code{TimeVectVariable}
\item \code{TimeVectArray}
\item \code{TimeVectLayerVariable}
\item \code{TimeVectLayerArray}
\item \code{TimeRastVariable}
\item \code{TimeRastArray}
\item \code{TimeRastLayerVariable}
\item \code{TimeRastLayerArray}
}}

\item{fn_}{(char) filenames}

\item{other_Attr}{(vector of named char) other global attributes}
}
\description{
Write \code{TimeSpatData} to nc-file.
}
