% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{crop_tsd}
\alias{crop_tsd}
\alias{crop_tsd.TimeVectVariable}
\alias{crop_tsd.TimeVectArray}
\alias{crop_tsd.TimeRastVariable}
\alias{crop_tsd.TimeRastArray}
\alias{crop_tsd.TimeRastLayerVariable}
\alias{crop_tsd.TimeRastLayerArray}
\title{crop TimeSpatData (Spat-Dimension)}
\usage{
crop_tsd(tsd_data, mask_area)

\method{crop_tsd}{TimeVectVariable}(tsd_data, mask_area)

\method{crop_tsd}{TimeVectArray}(tsd_data, mask_area)

\method{crop_tsd}{TimeRastVariable}(tsd_data, mask_area)

\method{crop_tsd}{TimeRastArray}(tsd_data, mask_area)

\method{crop_tsd}{TimeRastLayerVariable}(tsd_data, mask_area)

\method{crop_tsd}{TimeRastLayerArray}(tsd_data, mask_area)
}
\arguments{
\item{tsd_data}{\code{TimeSpatData} data
\itemize{
\item \code{TimeVectVariable}
\item \code{TimeVectArray}
\item \code{TimeRastVariable}
\item \code{TimeRastArray}
\item \code{TimeRastLayerVariable}
\item \code{TimeRastLayerArray}
}}

\item{mask_area}{(SpatVector) \strong{one} Polygon, from \code{\link[terra:vect]{terra::vect()}}}
}
\description{
Crop (mask) \code{TimeSpatData} new \code{TimeSpatData}.
Every time can only \strong{one Polygon} as the \code{mask_area}.
\itemize{
\item for \code{TimeVectData} the polygons, lines oder points will mit \code{mask_area} first intersect \link[terra:intersect]{terra::intersect}
then only the polygons, lines oder points which in the \code{mask_area} will be exported.
\item for \code{TimeRastData} or \code{TimeRastLayerData} the spatial dimension will first crop \link[terra:crop]{terra::crop} than \link[terra:mask]{terra::mask}
the raster-cells will be in NA setted, when they are not in the \code{mask_area}.
}
}
